Improve technical details
I've made several significant changes in this pull request.

First off the deliverables section has been dramatically expanded to
include a lot of the mundane details of operating a bridge. Please review
and comment if you feel this is too verbose.

Second I have changed one of the delivery timelines. Peggy V2 is slated for
January not December in the proposal document. I corrected that here.

Finally I have removed live upgrade language now that we have decided not to
go forward with that proposal. The clarification of live upgrades as distinct
from the status quo (chain halt upgrades) is no longer needed and is just
extra verbosity now that it's no longer on the table.